+++ src/classes/Modal.js
@@ -1,7 +1,10 @@
 import AbstractDomElement from './AbstractDomElement.js'
+import { ThrottledEvent } from 'oneloop.js'
+import uniqid from 'uniqid'
 
 /**
  * Modal Class
+ *
  * @author Milan Ricoul
  */
 class Modal extends AbstractDomElement {
@@ -17,6 +20,7 @@ class Modal extends AbstractDomElement {
 
     this.isOpened = false
     this.triggerButton = null
+    this._onResizeHandler = onResize.bind(this)
     this.close = this.close.bind(this)
     this.open = this.open.bind(this)
     this.handleButtonClick = this.handleButtonClick.bind(this)
@@ -30,12 +34,17 @@ class Modal extends AbstractDomElement {
       this._onClose = onClose.bind(this)
     }
 
-    this.init()
+    this.initialized = false
+
+    new ThrottledEvent(window, 'resize').add('resize', this._onResizeHandler)
+    this._onResizeHandler()
   }
 
   /**
    * Initialization
+   *
    * @returns {void}
+   *
    * @author Milan Ricoul
    */
   init() {
@@ -46,7 +55,7 @@ class Modal extends AbstractDomElement {
     if (el.id) {
       this.id = el.id
     } else {
-      this.id = this.defineId()
+      this.id = `modal-${uniqid()}`
 
       el.id = this.id
     }
@@ -64,13 +73,21 @@ class Modal extends AbstractDomElement {
     }
 
     // Set aria-controls attribute to close button
-    if (s.closeButtonSelector) {
+    if (s.closeButtonSelector && el.querySelector(s.closeButtonSelector)) {
       const closeButton = el.querySelector(s.closeButtonSelector)
+      closeButton.id = `${this.id}-close`
       closeButton.setAttribute('aria-controls', this.id)
       closeButton.addEventListener('click', this.close)
     }
 
-    document.querySelectorAll(`button[data-modal="${this.id}"]`).forEach((btn) => {
+    // if setting triggerButton is defined and exists, set aria-controls attribute to this button
+    if (s.triggerSelector && document.querySelectorAll(s.triggerSelector).length) {
+      document.querySelectorAll(s.triggerSelector).forEach((btn) => {
+        btn.setAttribute('aria-controls', this.id)
+      })
+    }
+
+    document.querySelectorAll(`button[aria-controls="${this.id}"]:not(#${this.id}-close)`).forEach((btn) => {
       btn.addEventListener('click', this.handleButtonClick)
     })
 
@@ -78,25 +95,12 @@ class Modal extends AbstractDomElement {
   }
 
   /**
-   * Set id to the modal dialog
-   * @returns {String}
-   * @author Milan Ricoul
-   */
-  defineId() {
-    const s = this._settings
-    let i = 1
-
-    while (document.getElementById(`${s.prefixId}-${i}`)) {
-      i++
-    }
-
-    return `${s.prefixId}-${i}`
-  }
-
-  /**
    * Handle button click
+   *
    * @param {MouseEvent} e click event handler
+   *
    * @returns {Void}
+   *
    * @author Milan Ricoul
    */
   handleButtonClick(e) {
@@ -107,7 +111,9 @@ class Modal extends AbstractDomElement {
 
   /**
    * Open modal
+   *
    * @returns {Void}
+   *
    * @author Milan Ricoul
    */
   open() {
@@ -123,7 +129,9 @@ class Modal extends AbstractDomElement {
 
   /**
    * Close modal
+   *
    * @returns {Void}
+   *
    * @author Milan Ricoul
    */
   close() {
@@ -141,8 +149,11 @@ class Modal extends AbstractDomElement {
 
   /**
    * Handle keydown event
+   *
    * @param {KeyboardEvent} e keydown event handler
+   *
    * @returns {Void}
+   *
    * @author Milan Ricoul
    */
   handleKeydown(e) {
@@ -159,8 +170,11 @@ class Modal extends AbstractDomElement {
 
   /**
    * Check if the next focusable element in dialog exists, else focus the first focusable element in dialog
+   *
    * @param {KeyboardEvent} e keyboard event handler
+   *
    * @returns {Void}
+   *
    * @author Milan Ricoul
    */
   checkNextFocusableElement(e) {
@@ -183,6 +197,7 @@ class Modal extends AbstractDomElement {
 
   /**
    * Destroy method
+   *
    * @author Milan Ricoul
    */
   destroy() {
@@ -199,13 +214,37 @@ class Modal extends AbstractDomElement {
   }
 }
 
+/**
+ * Events
+ *
+ * @returns {void}
+ *
+ * @author Milan Ricoul
+ */
+function onResize() {
+  const s = this._settings
+
+  if (!s.mediaQuery && !this.initialized) {
+    this.init()
+  }
+
+  if (s.mediaQuery && this.initialized && !s.mediaQuery.matches) {
+    this.destroy()
+  } else if (!this.initialized && s.mediaQuery && s.mediaQuery.matches) {
+    this.init()
+  } else if (!this.initialized && s.mediaQuery && !s.mediaQuery.matches && this._element.hasAttribute('style')) {
+    this._element.removeAttribute('style')
+  }
+}
+
 Modal.defaults = {
-  prefixId: 'dialog',
-  labelSelector: false,
+  closeButtonSelector: '.modal__close',
   descriptionSelector: false,
-  closeButtonSelector: false,
+  labelSelector: false,
+  mediaQuery: null,
   onOpen: null,
   onClose: null,
+  triggerSelector: false,
 }
 
 export default Modal