+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.3.4 - 2023-06-14
+
+- Update `@playwright/test`
+- Add `mediaQuery` option for Modal
+- Add uniqid for Modal
+- Add manual and automatic id for Modal
+
 ## 1.3.3 - 2023-06-13
 
 - Remove Slider presets